# Performance: access current span through its own field of Context, not the generic map