This would go in after #1232, but I need the new benchmark there to show the improvement here.  (Review the 2nd commit for now.)

## Changes

Add the `trace` field below:
```rust
pub struct Context {
    pub(super) trace: Option<Arc<super::trace::context::SynchronizedSpan>>,
    entries: HashMap<TypeId, Arc<dyn Any + Sync + Send>, BuildHasherDefault<IdHasher>>,
}
```
so that current span manipulation doesn't have to traverse the expensive HashMap and then get dispatched with virtual function calls.

Further, this keeps the HashMap empty for the common case of just having a span in context and no other types (like baggage).

I get that the opentelemetry specification is trying to be general with its view that `Context` is extensible and can hold anything, but there comes a real performance penalty for handling everything through that abstraction.

Here are preliminary results:
```
new_span/if_parent_sampled/in-cx/alt
                        time:   [443.41 ns 445.05 ns 447.27 ns]
                        thrpt:  [2.2358 Melem/s 2.2469 Melem/s 2.2553 Melem/s]
                 change:
                        time:   [-9.4406% -9.0621% -8.6385%] (p = 0.00 < 0.05)
                        thrpt:  [+9.4553% +9.9651% +10.425%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
new_span/if_recording/in-cx/alt
                        time:   [9.6986 ns 9.7092 ns 9.7208 ns]
                        thrpt:  [102.87 Melem/s 102.99 Melem/s 103.11 Melem/s]
                 change:
                        time:   [-33.193% -33.097% -33.006%] (p = 0.00 < 0.05)
                        thrpt:  [+49.266% +49.470% +49.685%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
new_span/if_parent_sampled/in-cx/spec
                        time:   [406.51 ns 407.03 ns 407.66 ns]
                        thrpt:  [2.4530 Melem/s 2.4568 Melem/s 2.4600 Melem/s]
                 change:
                        time:   [-12.614% -12.249% -11.821%] (p = 0.00 < 0.05)
                        thrpt:  [+13.406% +13.959% +14.435%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
new_span/if_recording/in-cx/spec
                        time:   [16.089 ns 16.116 ns 16.150 ns]
                        thrpt:  [61.920 Melem/s 62.050 Melem/s 62.154 Melem/s]
                 change:
                        time:   [-58.218% -58.148% -58.081%] (p = 0.00 < 0.05)
                        thrpt:  [+138.55% +138.94% +139.34%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
new_span/if_parent_sampled/no-cx/alt
                        time:   [9.0403 ns 9.0478 ns 9.0557 ns]
                        thrpt:  [110.43 Melem/s 110.52 Melem/s 110.62 Melem/s]
                 change:
                        time:   [-14.455% -14.352% -14.255%] (p = 0.00 < 0.05)
                        thrpt:  [+16.625% +16.758% +16.897%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
new_span/if_recording/no-cx/alt
                        time:   [9.5251 ns 9.5388 ns 9.5518 ns]
                        thrpt:  [104.69 Melem/s 104.83 Melem/s 104.99 Melem/s]
                 change:
                        time:   [-13.539% -13.332% -13.137%] (p = 0.00 < 0.05)
                        thrpt:  [+15.124% +15.383% +15.659%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
new_span/if_parent_sampled/no-cx/spec
                        time:   [182.64 ns 182.83 ns 183.08 ns]
                        thrpt:  [5.4621 Melem/s 5.4694 Melem/s 5.4753 Melem/s]
                 change:
                        time:   [-17.834% -16.546% -15.385%] (p = 0.00 < 0.05)
                        thrpt:  [+18.183% +19.826% +21.706%]
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe
new_span/if_recording/no-cx/spec
                        time:   [11.692 ns 11.713 ns 11.736 ns]
                        thrpt:  [85.205 Melem/s 85.374 Melem/s 85.527 Melem/s]
                 change:
                        time:   [+13.887% +14.200% +14.533%] (p = 0.00 < 0.05)
                        thrpt:  [-12.689% -12.434% -12.193%]
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
new_span/if_parent_sampled/no-sdk/alt
                        time:   [8.9813 ns 8.9963 ns 9.0120 ns]
                        thrpt:  [110.96 Melem/s 111.16 Melem/s 111.34 Melem/s]
                 change:
                        time:   [-14.718% -14.566% -14.413%] (p = 0.00 < 0.05)
                        thrpt:  [+16.840% +17.050% +17.258%]
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  8 (8.00%) low mild
  5 (5.00%) high mild
new_span/if_recording/no-sdk/alt
                        time:   [9.5512 ns 9.5583 ns 9.5660 ns]
                        thrpt:  [104.54 Melem/s 104.62 Melem/s 104.70 Melem/s]
                 change:
                        time:   [-13.348% -13.168% -13.003%] (p = 0.00 < 0.05)
                        thrpt:  [+14.947% +15.165% +15.404%]
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  5 (5.00%) high mild
  2 (2.00%) high severe
new_span/if_parent_sampled/no-sdk/spec
                        time:   [83.663 ns 83.759 ns 83.879 ns]
                        thrpt:  [11.922 Melem/s 11.939 Melem/s 11.953 Melem/s]
                 change:
                        time:   [-23.079% -22.948% -22.795%] (p = 0.00 < 0.05)
                        thrpt:  [+29.525% +29.782% +30.003%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
new_span/if_recording/no-sdk/spec
                        time:   [11.606 ns 11.617 ns 11.627 ns]
                        thrpt:  [86.010 Melem/s 86.084 Melem/s 86.163 Melem/s]
                 change:
                        time:   [+12.821% +12.973% +13.103%] (p = 0.00 < 0.05)
                        thrpt:  [-11.585% -11.483% -11.364%]
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) low mild
  2 (2.00%) high mild

```

## Merge requirement checklist

* [x] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [x] Unit tests added/updated (if applicable)
* [x] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [x] Changes in public API reviewed (if applicable)